+++ NEWS.md
@@ -1,5 +1,8 @@
 # devtools 1.9.1.9000
 
+* `release()` and `build_win()` use `--compact-vignettes=both`
+   to avoid WARNINGs with the built package (#981, @krlmlr).
+
 * 'Check failed:' now includes the package name for when Ncpus>1 so you
    know which package has failed and can start looking at the output without
    needing to wait for all packages to finish (@mattdowle).